Updated json schema
Decided to change the json schema that is generated. Rather than doing individual records for each of the games streamed, I create a recrod for the streaming session. This way that all of the data stays together for that stream session and can be used when generating the web page.